Standardize test and input file names
Generally, there is a single sample input and single main input. Question 1 was an exception in that there was two different sample inputs, but I think we can handle this manually when that case arises.

I propose the following convention for test names: `<programmer>_q<day>_p<1|2>_<main|sample>`

This would allow us to get as much granularity as we need:

* `cargo test` - Run all tests
* `cargo test roar` - Run all roar's tests
* `cargo test gee_q02` - Run all gee's tests for question 2
* `cargo test roar_q03_p1` - Run the sample and main tests for roar, question 3, part 1
* `cargo test gee_q01_p2_main` - Run only gee's question 1, part 2, main test

It would also work if you wanted the following:

* `cargo test q03` - Run both gee's and roar's tests for question 3